+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 3.0.1 under development
 
-- New #139: Add `ArrayHelper::restrictedMerge()` method that allow to merge two or more arrays recursively with
+- New #139: Add `ArrayHelper::parametrizedMerge()` method that allow to merge two or more arrays recursively with
   specified depth (@vjik)
 
 ## 3.0.0 January 12, 2023